Move the graphql datastructures to a directory. This should make no difference to the public API as these datastructures are already exported from the `graphql` module. This also splits out the visitor from the response file.
