+++ apollo-router/src/graphql/fixtures/federated_ships_deferred_query.graphql
@@ -1,0 +1,8 @@
+{
+    ships {
+        name
+        ... @defer {
+            registrationFee
+        }
+    }
+}
