+++ apollo-router/src/graphql/fixtures/federated_ships_deferred_response.json
@@ -1,0 +1,12 @@
+{
+    "data": {
+        "ships": [
+            {
+                "name": "Boaty McBoatface"
+            },
+            {
+                "name": "HMS Grapherson"
+            }
+        ]
+    }
+}
