+++ apollo-router/src/graphql/fixtures/federated_ships_inline_fragment_query.graphql
@@ -1,0 +1,16 @@
+{
+    ships {
+        owner {
+            ... on User {
+                licenseNumber
+                name
+            }
+        }
+    }
+    users {
+        ... on User {
+            licenseNumber
+            name
+        }
+    }
+}
