+++ apollo-router/src/graphql/fixtures/federated_ships_named_query.graphql
@@ -1,0 +1,6 @@
+query NamedQuery {
+    users {
+        licenseNumber
+        name
+    }
+}
