+++ apollo-router/src/graphql/fixtures/federated_ships_named_response.json
@@ -1,0 +1,14 @@
+{
+    "data": {
+        "users": [
+            {
+                "licenseNumber": 1,
+                "name": "Kate Chopin"
+            },
+            {
+                "licenseNumber": 2,
+                "name": "Paul Auster"
+            }
+        ]
+    }
+}
