+++ apollo-router/src/graphql/fixtures/federated_ships_required_query.graphql
@@ -1,0 +1,6 @@
+{
+    ships {
+        name
+        registrationFee
+    }
+}
