+++ apollo-router/src/graphql/fixtures/federated_ships_required_response.json
@@ -1,0 +1,14 @@
+{
+    "data": {
+        "ships": [
+            {
+                "name": "Boaty McBoatface",
+                "registrationFee": 200
+            },
+            {
+                "name": "HMS Grapherson",
+                "registrationFee": 150
+            }
+        ]
+    }
+}
