+++ apollo-router/src/graphql/fixtures/federated_ships_schema.graphql
@@ -1,0 +1,75 @@
+schema
+  @link(url: "https://specs.apollo.dev/link/v1.0")
+  @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
+{
+  query: Query
+}
+
+directive @join__enumValue(graph: join__Graph!) repeatable on ENUM_VALUE
+
+directive @join__field(graph: join__Graph, requires: join__FieldSet, provides: join__FieldSet, type: String, external: Boolean, override: String, usedOverridden: Boolean) repeatable on FIELD_DEFINITION | INPUT_FIELD_DEFINITION
+
+directive @join__graph(name: String!, url: String!) on ENUM_VALUE
+
+directive @join__implements(graph: join__Graph!, interface: String!) repeatable on OBJECT | INTERFACE
+
+directive @join__type(graph: join__Graph!, key: join__FieldSet, extension: Boolean! = false, resolvable: Boolean! = true, isInterfaceObject: Boolean! = false) repeatable on OBJECT | INTERFACE | UNION | ENUM | INPUT_OBJECT | SCALAR
+
+directive @join__unionMember(graph: join__Graph!, member: String!) repeatable on UNION
+
+directive @link(url: String, as: String, for: link__Purpose, import: [link__Import]) repeatable on SCHEMA
+
+type Address
+  @join__type(graph: USERS)
+  @join__type(graph: VEHICLES)
+{
+  street: String! @join__field(graph: USERS)
+  zipCode: Int! @join__field(graph: USERS) @join__field(graph: VEHICLES, external: true)
+}
+
+scalar join__FieldSet
+
+enum join__Graph {
+  USERS @join__graph(name: "users", url: "http://localhost:4001")
+  VEHICLES @join__graph(name: "vehicles", url: "http://localhost:4002")
+}
+
+scalar link__Import
+
+enum link__Purpose {
+  """
+  `SECURITY` features provide metadata necessary to securely resolve fields.
+  """
+  SECURITY
+
+  """
+  `EXECUTION` features provide metadata necessary for operation execution.
+  """
+  EXECUTION
+}
+
+type Query
+  @join__type(graph: USERS)
+  @join__type(graph: VEHICLES)
+{
+  users: [User!]! @join__field(graph: USERS)
+  ships: [Ship!]! @join__field(graph: VEHICLES)
+}
+
+type Ship
+  @join__type(graph: VEHICLES, key: "id")
+{
+  id: ID!
+  name: String!
+  owner: User!
+  registrationFee: Float @join__field(graph: VEHICLES, requires: "owner { addresses { zipCode } }")
+}
+
+type User
+  @join__type(graph: USERS, key: "licenseNumber")
+  @join__type(graph: VEHICLES, key: "licenseNumber")
+{
+  licenseNumber: ID!
+  name: String! @join__field(graph: USERS)
+  addresses: [Address!]! @join__field(graph: USERS) @join__field(graph: VEHICLES, external: true)
+}
